requirements.txt
PyQt5>=5.15.0
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.3.0
Pillow>=8.0.0
